v1.2.3 split of nodes in separate files + readme update
