+++ sap_hana_config.html
@@ -1,0 +1,35 @@
+<script type="text/x-red" data-template-name="sap_hana_config">
+    <div class="form-row">
+        <label for="node-config-input-host"><i class="fa fa-globe"></i> Host</label>
+        <input type="text" id="node-config-input-host" placeholder="Database Server Name">
+    </div>
+    <div class="form-row">
+        <label for="node-config-input-port"><i class="fa fa-globe"></i> Port</label>
+        <input type="text" id="node-config-input-port" placeholder="Database Server Port">
+    </div>
+    <div class="form-row">
+        <label for="node-config-input-user"><i class="fa fa-user"></i> User</label>
+        <input type="text" id="node-config-input-user" placeholder="Username">
+    </div>
+        <div class="form-row">
+        <label for="node-config-input-password"><i class="fa fa-lock"></i> Password</label>
+        <input type="password" id="node-config-input-password" placeholder="Password">
+    </div>
+</script>
+<script type="text/javascript">
+    RED.nodes.registerType('sap_hana_config',{
+        category: 'config',
+        defaults: {
+            host: {value:"hxehost",required:true},
+            port: {value:"39015", required:true},
+			user: {value:"",required:true},
+            password: {value:"",required:true}			
+        },
+        label: function() {
+            if (this.name)
+                return this.name;
+            else
+                return "sap_hana_config";
+        }
+    });
+</script>
