--- saphana.html
@@ -1,145 +1,0 @@
-<script type="text/x-red" data-template-name="sap_hana_config">
-    <div class="form-row">
-        <label for="node-config-input-host"><i class="fa fa-globe"></i> Host</label>
-        <input type="text" id="node-config-input-host" placeholder="Database Server Name">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-port"><i class="fa fa-globe"></i> Port</label>
-        <input type="text" id="node-config-input-port" placeholder="Database Server Port">
-    </div>
-    <div class="form-row">
-        <label for="node-config-input-user"><i class="fa fa-user"></i> User</label>
-        <input type="text" id="node-config-input-user" placeholder="Username">
-    </div>
-        <div class="form-row">
-        <label for="node-config-input-password"><i class="fa fa-lock"></i> Password</label>
-        <input type="password" id="node-config-input-password" placeholder="Password">
-    </div>
-</script>
-<script type="text/javascript">
-    RED.nodes.registerType('sap_hana_config',{
-        category: 'config',
-        defaults: {
-            host: {value:"hxehost",required:true},
-            port: {value:"39015", required:true},
-			user: {value:"",required:true},
-            password: {value:"",required:true}			
-        },
-        label: function() {
-            if (this.name)
-                return this.name;
-            else
-                return "sap_hana_config";
-        }
-    });
-</script>
-<script type="text/x-red" data-template-name="run_query">
-    <div class="form-row">
-        <label for="node-input-mydb"><i class="fa fa-database"></i> Database</label>
-        <input type="text" id="node-input-mydb">
-    </div>
-    <div class="form-row">
-        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-</script>
-<script type="text/javascript">
-    RED.nodes.registerType('run_query',{
-        category: 'sap_hana',
-        color:"#e97b00",
-        defaults: {
-            mydb: {type:"sap_hana_config", required:true},
-            name: {value:"runquery1", required:true}
-        },
-        inputs:1,
-        outputs:1,
-        icon: "db.png",
-        label: function() {
-            var levelNode = RED.nodes.node(this.mydb);
-            return this.name||(levelNode?levelNode.label():"run_query");
-        },
-            labelStyle: function() {
-            return this.name?"node_label_italic":"";
-        }
-    });
-</script>
-<script type="text/x-red" data-template-name="table_monitor">
-    <div class="form-row">
-        <label for="node-input-mydb"><i class="fa fa-database"></i> Database</label>
-        <input type="text" id="node-input-mydb">
-    </div>
-    <div class="form-row">
-        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-    <div class="form-row">
-        <label for="node-input-query"><i class="fa fa-tag"></i> Query</label>
-        <input type="text" id="node-input-query" placeholder="Query">
-    </div>	
-    <div class="form-row">
-        <label for="node-input-interval"><i class="fa fa-tag"></i> Interval</label>
-        <input type="text" id="node-input-interval" placeholder="in seconds">
-    </div>	
-</script>
-<script type="text/javascript">
-    RED.nodes.registerType('table_monitor',{
-        category: 'sap_hana',
-        color:"#e97b00",
-        defaults: {
-            mydb: {type:"sap_hana_config", required:true},
-            name: {value:"tablemonitor1", required:true},
-			query: {value:"SELECT COUNT(*) FROM DUMMY", required:true},
-			interval : {value:"1", required:true}
-        },
-        inputs: 0,
-        outputs: 1,
-        icon: "db.png",
-        label: function() {
-            var levelNode = RED.nodes.node(this.mydb);
-            return this.name || (levelNode ? levelNode.label() : "table_monitor");
-        },
-            labelStyle: function() {
-            return this.name?"node_label_italic":"";
-        }
-    });
-</script>
-<script type="text/x-red" data-template-name="table_insert">
-    <div class="form-row">
-        <label for="node-input-mydb"><i class="fa fa-database"></i> Database</label>
-        <input type="text" id="node-input-mydb">
-    </div>
-    <div class="form-row">
-        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
-        <input type="text" id="node-input-name" placeholder="Name">
-    </div>
-    <div class="form-row">
-        <label for="node-input-table"><i class="fa fa-tag"></i> Table</label>
-        <input type="text" id="node-input-table" placeholder="Table">
-    </div>	
-    <div class="form-row">
-        <label for="node-input-columns"><i class="fa fa-tag"></i> Columns</label>
-        <input type="text" id="node-input-columns" placeholder="Column1,Column2">
-    </div>	
-</script>
-<script type="text/javascript">
-    RED.nodes.registerType('table_insert',{
-        category: 'sap_hana',
-        color:"#e97b00",
-        defaults: {
-            mydb: {type:"sap_hana_config", required:true},
-            name: {value:"tableinsert1", required:true},
-			table: {value:"table1", required:true},
-			columns: {value:"column1", required:true}
-        },
-        inputs: 1,
-        outputs: 1,
-        icon: "db.png",
-        label: function() {
-            var levelNode = RED.nodes.node(this.mydb-insert);
-            return this.name || (levelNode ? levelNode.label() : "table_insert");
-        },
-            labelStyle: function() {
-            return this.name?"node_label_italic":"";
-        }
-    });
-</script>
