+++ table_insert.html
@@ -1,0 +1,40 @@
+<script type="text/x-red" data-template-name="table_insert">
+    <div class="form-row">
+        <label for="node-input-mydb"><i class="fa fa-database"></i> Database</label>
+        <input type="text" id="node-input-mydb">
+    </div>
+    <div class="form-row">
+        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <div class="form-row">
+        <label for="node-input-table"><i class="fa fa-tag"></i> Table</label>
+        <input type="text" id="node-input-table" placeholder="Table">
+    </div>	
+    <div class="form-row">
+        <label for="node-input-columns"><i class="fa fa-tag"></i> Columns</label>
+        <input type="text" id="node-input-columns" placeholder="Column1,Column2">
+    </div>	
+</script>
+<script type="text/javascript">
+    RED.nodes.registerType('table_insert',{
+        category: 'sap_hana',
+        color:"#e97b00",
+        defaults: {
+            mydb: {type:"sap_hana_config", required:true},
+            name: {value:"tableinsert1", required:true},
+			table: {value:"table1", required:true},
+			columns: {value:"column1", required:true}
+        },
+        inputs: 1,
+        outputs: 1,
+        icon: "db.png",
+        label: function() {
+            var levelNode = RED.nodes.node(this.mydb-insert);
+            return this.name || (levelNode ? levelNode.label() : "table_insert");
+        },
+            labelStyle: function() {
+            return this.name?"node_label_italic":"";
+        }
+    });
+</script>
