+++ table_monitor.html
@@ -1,0 +1,40 @@
+<script type="text/x-red" data-template-name="table_monitor">
+    <div class="form-row">
+        <label for="node-input-mydb"><i class="fa fa-database"></i> Database</label>
+        <input type="text" id="node-input-mydb">
+    </div>
+    <div class="form-row">
+        <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
+        <input type="text" id="node-input-name" placeholder="Name">
+    </div>
+    <div class="form-row">
+        <label for="node-input-query"><i class="fa fa-tag"></i> Query</label>
+        <input type="text" id="node-input-query" placeholder="Query">
+    </div>	
+    <div class="form-row">
+        <label for="node-input-interval"><i class="fa fa-tag"></i> Interval</label>
+        <input type="text" id="node-input-interval" placeholder="in seconds">
+    </div>	
+</script>
+<script type="text/javascript">
+    RED.nodes.registerType('table_monitor',{
+        category: 'sap_hana',
+        color:"#e97b00",
+        defaults: {
+            mydb: {type:"sap_hana_config", required:true},
+            name: {value:"tablemonitor1", required:true},
+			query: {value:"SELECT COUNT(*) FROM DUMMY", required:true},
+			interval : {value:"1", required:true}
+        },
+        inputs: 0,
+        outputs: 1,
+        icon: "db.png",
+        label: function() {
+            var levelNode = RED.nodes.node(this.mydb);
+            return this.name || (levelNode ? levelNode.label() : "table_monitor");
+        },
+            labelStyle: function() {
+            return this.name?"node_label_italic":"";
+        }
+    });
+</script>
